requirements.txt
absl-py==0.2.0
alabaster==0.7.10
arrow==0.12.1
asn1crypto==0.24.0
astor==0.6.2
astroid==1.6.3
autopep8==1.3.5
Babel==2.5.3
backcall==0.1.0
binaryornot==0.4.4
bleach==1.5.0
bokeh==0.12.15
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
cryptography==2.1.4
cycler==0.10.0
Cython==0.28.2
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
flake8==3.5.0
flexx==0.4.1
future==0.16.0
gast==0.2.0
grpcio==1.11.0
h5py==2.8.0rc1
html5lib==0.9999999
idna==2.6
imageio==2.3.0
imagesize==1.0.0
imgaug==0.2.5
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
jedi==0.12.0
Jinja2==2.10
jinja2-time==0.2.0
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.5
-e git+git@gitlab-vs.informatik.uni-ulm.de:doj49/hunse_kitti_toolkit.git@8f952a08272d7560a9ada03066c98e2aa4081aba#egg=kitti
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mask-rcnn==2.1
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nose==1.3.7
notebook==5.4.1
numpy==1.14.3
olefile==0.45.1
opencv-python==3.4.0.12
packaging==17.1
pandocfilters==1.4.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
pkg-config==0.0.1
poyo==0.4.1
pprint==0.1
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pycocotools==2.0
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pyflakes==1.6.0
pygame==1.9.3
Pygments==2.2.0
pylint==1.8.4
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.7.2
python-dotenv==0.8.2
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
rope==0.10.7
ruamel.yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.7.4
sphinx-rtd-theme==0.3.1
sphinxcontrib-websupport==1.0.1
tensorboard==1.8.0
tensorflow-gpu==1.8.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
tqdm==4.23.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
whichcraft==0.4.1
widgetsnbextension==3.2.1
wrapt==1.10.11
